% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents_studies.R
\name{combine_agents_studies}
\alias{combine_agents_studies}
\title{Combine Agents and Studies into one data.frame}
\usage{
combine_agents_studies(agents, studies)
}
\arguments{
\item{agents}{(\code{data.frame})\cr
A \code{data.frame} built with \code{\link[=create_agents]{create_agents()}} containing the agents' properties.}

\item{studies}{(\code{data.frame})\cr
A \code{data.frame} built with \code{\link[=create_studies]{create_studies()}} containing the studies' properties.}
}
\value{
(\code{data.frame})
A \code{data.frame} containing the combined properties of agents and studies.
The \code{data.frame} is created by merging the two input \code{data.frame}s by the \code{agent_id} column,
i.e. agents' properties are added to the studies' properties based on the \code{agent_id} assigned to that study.
}
\description{
Function to create a \code{data.frame} containing the combined properties of agents and studies, refered to as \code{complete_studies}.
}
\examples{
# Create agents and studies
agents <- create_agents(
  sbj_prob_alpha = 2, sbj_prob_beta = 3,
  sbj_effect_mu = 0.3, sbj_effect_sigma2 = 1.2,
  sbj_error_mu = 0.6, sbj_error_kappa = 1,
  sbj_error_var_alpha = 3, sbj_error_var_beta = 2,
  agent_id = "Alice"
)
studies <- create_studies(
  N = 30, resources = 100, cost = 4, benefit = 3,
  obj_prob_fault = 0.4,
  obj_effect_mu = 0.4, obj_effect_sigma2 = 0.5,
  obj_error_size_mu = 0.1, obj_error_size_sigma2 = 0.2,
  study_id = "Alice2022", agent_id = "Alice"
)
# Combine agents and studies
combine_agents_studies(agents, studies)
}
