% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_distributions_heatmap}
\alias{plot_distributions_heatmap}
\title{Plot Probability Distributions over Rounds as Heatmap}
\usage{
plot_distributions_heatmap(literature, study_id, type)
}
\arguments{
\item{literature}{(\code{list()})\cr
A list containing the simulated literature, generated by \code{\link[=simulate_literature]{simulate_literature()}}.}

\item{study_id}{(\code{character()} or \code{integer()})\cr
A unique identifier of a study indicating for which study the plot should be generated.}

\item{type}{(\code{character()})\cr
a character string indicating the type of distribution to plot. Can be either "after_effect" or "after_fault_ind".}
}
\value{
A \code{ggplot} object that can be further customized using the \code{ggplot2} package.
}
\description{
Plots the discrete probability distributions simulated through \code{\link[=simulate_literature]{simulate_literature()}} per round as a heatmap.
}
