% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_fault_beliefs}
\alias{plot_fault_beliefs}
\title{Plot Belief for Finding a Fault across Rounds}
\usage{
plot_fault_beliefs(literature)
}
\arguments{
\item{literature}{(\code{list})\cr
The simulated literature data, generated by \code{\link[=simulate_literature]{simulate_literature()}}.}
}
\value{
A \code{list} of \code{ggplot} objects, one per study.
}
\description{
Plot Belief for Finding a Fault across Rounds
}
