% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_histogram_n_faults}
\alias{plot_histogram_n_faults}
\title{Plot Histogram of Number of Remaining Faults}
\usage{
plot_histogram_n_faults(literature, study_id = NULL, type = "relative")
}
\arguments{
\item{literature}{(\code{list()})\cr
A list containing the simulated literature, generated by \code{\link[=simulate_literature]{simulate_literature()}}.}

\item{study_id}{(\code{character()} or \code{integer()})\cr
A vector of (possibly multiple) unique identifiers of studies indicating for which studies the plot should be generated.
If \code{NULL}, all studies in \code{literature} are plotted. Default is \code{NULL}.}

\item{type}{(\code{character()})\cr
a character string indicating whether to plot the absolute number of undiscovered faults or the number of undiscovered
faults relative to the total number of faults that existed in the study. Can be either \code{"absolute"} or \code{"relative"}.
Default is \code{"relative"}.}
}
\value{
A \code{ggplot} object that can be further customized using the \code{ggplot2} package.
}
\description{
Plots the distribution of the number of remaining faults across the specified studies as a histogram.
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
library(ggplot2)

# Create study and agent
agent <- create_agents(
  agent_id = seq(1, 20),
  subj_effect_mu = 0.3, subj_effect_sigma = 1.2,
  subj_prob_fault_alpha = 0.2, subj_prob_fault_beta = 0.4,
  subj_error_size_mu = 0.6, subj_error_size_sigma = 0.2
)
study <- create_studies(
  agent_id = seq(1, 20),
  N = 30, resources = 1000, cost = 400, benefit = 20,
  obj_effect_mu = 0.4, obj_effect_sigma = 0.5,
  obj_prob_fault = 0.4,
  obj_error_size_mu = 0.1, obj_error_size_sigma = 0.2
)
# Simulate the study
lit <- simulate_literature(agents = agent, studies = study, seed = 123)
# Plot the number of remaining faults per round
plot_histogram_n_faults(lit)
}

}
