% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_observed_effect_sizes}
\alias{plot_observed_effect_sizes}
\title{Plot Observed Effect Sizes across Rounds}
\usage{
plot_observed_effect_sizes(literature)
}
\arguments{
\item{literature}{(\code{list})\cr
The simulated literature data, generated by \code{\link[=simulate_literature]{simulate_literature()}}.}
}
\value{
A \code{list} of \code{ggplot} objects, one per study.
}
\description{
Plot Observed Effect Sizes across Rounds
}
