% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_posteriors}
\alias{plot_posteriors}
\title{Plot Posterior for Number of Faults across Rounds as Heatmap}
\usage{
plot_posteriors(literature)
}
\arguments{
\item{literature}{(\code{list})\cr
The simulated literature data, generated by \code{\link[=simulate_literature]{simulate_literature()}}.}
}
\value{
A \code{list} of \code{ggplot} objects, one per study.
}
\description{
Plots the discrete probability distributions simulated through \code{\link[=simulate_literature]{simulate_literature()}} per round as a heatmap.
}
\examples{
# only want a few? index literature

}
