% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_rem_faults}
\alias{plot_rem_faults}
\title{Plot Number of Remaining Faults per Round}
\usage{
plot_rem_faults(literature, study_id = NULL)
}
\arguments{
\item{literature}{(\code{list()})\cr
A list containing the simulated literature, generated by \code{\link[=simulate_literature]{simulate_literature()}}.}

\item{study_id}{(\code{character()} or \code{integer()})\cr
A vector of (possibly multiple) unique identifiers of studies indicating for which studies the plot should be generated.
If \code{NULL}, all studies in \code{literature} are plotted. Default is \code{NULL}.}
}
\value{
A \code{ggplot} object that can be further customized using the \code{ggplot2} package.
}
\description{
Plot Number of Remaining Faults per Round
}
