% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_study_panels}
\alias{plot_study_panels}
\title{Plot Summary Panel per Study}
\usage{
plot_study_panels(literature)
}
\arguments{
\item{literature}{(\code{list})\cr
The simulated literature data, generated by \code{\link[=simulate_literature]{simulate_literature()}}.}
}
\value{
A \code{list} of \code{patchwork} objects, one per study.
}
\description{
Plot Summary Panel per Study
}
