% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_summary_panel}
\alias{plot_summary_panel}
\title{Plot Summary Panel}
\usage{
plot_summary_panel(literature, study_id = NULL)
}
\arguments{
\item{literature}{(\code{list()})\cr
A list containing the simulated literature, generated by \code{\link[=simulate_literature]{simulate_literature()}}.}

\item{study_id}{(\code{character()} or \code{integer()})\cr
A vector of (possibly multiple) unique identifiers of studies indicating for which studies the plot should be generated.
If \code{NULL}, all studies in \code{literature} are plotted. Default is \code{NULL}.}
}
\value{
\code{patchwork} / \code{gg} object
}
\description{
Plots a panel of three plots: the probability of finding a fault per round,
the number of remaining faults per round, and the observed effect sizes per round.
}
