% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scerrModel-package.R
\docType{package}
\name{scerrModel-package}
\alias{scerrModel}
\alias{scerrModel-package}
\title{scerrModel: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Keno Mersmann \email{keno.mersmann@campus.lmu.de}

}
\keyword{internal}
