% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{simulation_summary}
\alias{simulation_summary}
\title{Summarize Simulated Literature Data}
\usage{
simulation_summary(literature, simple = TRUE)
}
\arguments{
\item{literature}{(\code{list})\cr
The simulated literature data, generated by \code{\link[=simulate_literature]{simulate_literature()}}.}

\item{simple}{(\code{logical()})\cr
If \code{TRUE}, the output will be simplified to only contain a selection of key columns.
If \code{FALSE} the output will also contain all columns of \code{complete_studies} passed to / created by \code{\link[=simulate_literature]{simulate_literature()}}.
Default is \code{TRUE}.}
}
\value{
(\code{data.frame})\cr
A \code{data.frame} giving an overview over the simulated literature, with the following columns:
\itemize{
\item \code{study_id} (\code{integer()} or \code{character()})\cr
The unique identifier of the study.
\item \code{agent_id} (\code{integer()} or \code{character()})\cr
The unique identifier of the agent who performed the data analysis.
\item \code{stopped_in_round} (\code{integer()})\cr
Number of rounds the agent searched for faults.
\item \code{stopping_reason} (\code{character()})\cr
Reason for stopping the search. Can be "Search completed", "Expected utility too low" or "Resources depleted".
\item \code{final_resources} (\code{numeric()})\cr
Resources remaining when the search stopped.
\item \code{n_faults_total} (\code{integer()})\cr
Total number of faults in the study.
\item \code{n_faults_discovered} (\code{integer()})\cr
Number of faults discovered by the agent.
\item \code{true_effect} (\code{numeric()})\cr
True effect size of the study in the objective reality.
\item \code{init_obs_effect} (\code{numeric()})\cr
Initial observed effect size before any faults are corrected.
\item \code{final_obs_effect} (\code{numeric()})\cr
Final observed effect size when the agent stopped searching.
If \code{simple = FALSE}, the returned \code{data.frame} also contains all input parameters.
}
}
\description{
This function summarizes the simulated literature returned by \code{\link[=simulate_literature]{simulate_literature()}}.
}
\examples{
\dontrun{
# Create study and agent
agent <- create_agents(

)
study <- create_studies(

)

# Simulate the study
lit <- simulate_literature(agents = agent, studies = study, seed = 123)

# Summarize the literature (simple = TRUE)
simulation_summary(lit)

# Extensive Summary with all input parameters
ext_summary <- simulation_summary(lit, simple = FALSE)

# You can then index the information you need
}

}
