% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.R
\name{combine_agents_studies}
\alias{combine_agents_studies}
\title{Combine Agents and Studies into one data.frame}
\usage{
combine_agents_studies(agents, studies)
}
\arguments{
\item{agents}{(\code{data.frame})\cr
A \code{data.frame} built with \code{\link[=create_agents]{create_agents()}} containing the agents' properties.}

\item{studies}{(\code{data.frame})\cr
A \code{data.frame} built with \code{\link[=create_studies]{create_studies()}} containing the studies' properties.}
}
\value{
(\code{data.frame})
A \code{data.frame} containing the combined properties of agents and studies.
The \code{data.frame} is created by merging the two input \code{data.frame}s by the \code{agent_id} column,
i.e. agents' properties are added to the studies' properties based on the \code{agent_id} assigned to that study.
}
\description{
Function to create a \code{data.frame} containing the combined properties of agents and studies, refered to as \code{complete_studies}.
}
\examples{
# Create agents and studies
agents <- create_agents(
  agent_id = "Alice",
  subj_effect_mu = 0.3, subj_effect_sigma = 1.2,
  subj_prob_fault_alpha = 0.2, subj_prob_fault_beta = 0.4,
  subj_error_size_mu = 0.6, subj_error_size_sigma = 0.2
)
studies <- create_studies(
  study_id = "Alice2022", agent_id = "Alice",
  N = 30, resources = 100, cost = 4, benefit = 3,
  obj_effect_mu = 0.4, obj_effect_sigma = 0.5,
  obj_prob_fault = 0.4,
  obj_error_size_mu = 0.1, obj_error_size_sigma = 0.2
)
# Combine agents and studies
combine_agents_studies(agents, studies)
}
