% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents_studies.R
\name{create_agents}
\alias{create_agents}
\title{Create data.frame of Agents with all Required Properties}
\usage{
create_agents(
  sbj_prob_alpha,
  sbj_prob_beta,
  sbj_effect_mu,
  sbj_effect_sigma2,
  sbj_error_mu,
  sbj_error_kappa,
  sbj_error_var_alpha,
  sbj_error_var_beta,
  agent_id = NULL
)
}
\arguments{
\item{sbj_prob_alpha}{(\code{numeric()})\cr
The alpha parameter of the beta distribution for the probability of finding a fault in a search round. Must be greater than zero.}

\item{sbj_prob_beta}{(\code{numeric()})\cr
The beta parameter of the beta distribution for the probability of finding a fault in a search round. Must be greater than zero.}

\item{sbj_effect_mu}{(\code{numeric()})\cr
The mean of the agent's normal prior for the true effect.}

\item{sbj_effect_sigma2}{(\code{numeric()})\cr
The variance of the agent's normal prior for the true effect. Must be greater than zero.}

\item{sbj_error_mu}{(\code{numeric()})\cr
The mean parameter of the agent's Normal-Inverse-Gamma belief for error sizes.}

\item{sbj_error_kappa}{(\code{numeric()})\cr
The strength parameter of the agent's Normal-Inverse-Gamma belief for error sizes. Must be greater than zero.}

\item{sbj_error_var_alpha}{(\code{numeric()})\cr
The shape parameter of the agent's Inverse-Gamma belief for the error variance. Must be greater than 1.}

\item{sbj_error_var_beta}{(\code{numeric()})\cr
The scale parameter of the agent's Inverse-Gamma belief for the error variance. Must be greater than zero.}

\item{agent_id}{(\code{character()} or \code{integer()})\cr
Identifiers for Agents, has to be \strong{unique}, i.e. the same \code{agent_id} may only be used once.\cr
If \code{NULL}, IDs are generated as \code{a1}, \code{a2}, and so on. Default is \code{NULL}.}
}
\value{
(\code{data.frame})\cr
Agents \code{data.frame} with 9 columns and as many rows as the longest vector passed as an argument
to this function. Column names are standardized.
}
\description{
Generates a \code{data.frame} to define Agents by their properties in a standardized format.
The output can be combined with the output of \code{\link[=create_studies]{create_studies()}} to define a \link[=combine_agents_studies]{complete study}.
}
\details{
As a convenience functionality, this function can take atomic vectors of equal length or length 1 as input. If the input is of length 1,
the value is recycled to match the length of the longest input vector.
For more complex automatically generated inputs, users may want to use functions such as \code{\link[=rep]{rep()}} or \code{\link[=seq]{seq()}}.
}
\examples{
# Generate one agent, ID is generated automatically
create_agents(
  sbj_prob_alpha = 2, sbj_prob_beta = 3,
  sbj_effect_mu = 0.3, sbj_effect_sigma2 = 1.2,
  sbj_error_mu = 0.6, sbj_error_kappa = 1,
  sbj_error_var_alpha = 3, sbj_error_var_beta = 2
)
# Generate multiple agents
create_agents(
  sbj_prob_alpha = c(2, 3), sbj_prob_beta = c(3, 4),
  sbj_effect_mu = c(0.3, 0.4), sbj_effect_sigma2 = c(1.2, 1.3),
  sbj_error_mu = c(0.6, 0.7), sbj_error_kappa = 1,
  sbj_error_var_alpha = 3, sbj_error_var_beta = 2,
  agent_id = c("Alice", "Bob")
)

}
