% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents.R
\name{create_agents}
\alias{create_agents}
\title{Create data.frame of Agents with all Required Properties}
\usage{
create_agents(
  subj_effect_mu,
  subj_effect_sigma,
  subj_prob_fault_alpha,
  subj_prob_fault_beta,
  subj_error_size_mu,
  subj_error_size_sigma,
  agent_id = NULL
)
}
\arguments{
\item{subj_effect_mu}{(\code{numeric()})\cr
The mean of the agents' initial belief for the effect size distribution (normal).}

\item{subj_effect_sigma}{(\code{numeric()})\cr
The variance of the agents' initial belief for the effect size distribution (normal).}

\item{subj_prob_fault_alpha}{(\code{numeric()})\cr
The alpha parameter of the beta distribution for the probability of making an error in a search round.}

\item{subj_prob_fault_beta}{(\code{numeric()})\cr
The beta parameter of the beta distribution for the probability of making an error in a search round.}

\item{subj_error_size_mu}{(\code{numeric()})\cr
The mean of the agents' initial belief for the error size distribution (normal).}

\item{subj_error_size_sigma}{(\code{numeric()})\cr
The variance of the agents' initial belief for the error size distribution (normal).}

\item{agent_id}{(\code{character()} or \code{integer()})\cr
Identifiers for Agents, has to be \strong{unique}, i.e. the same \code{agent_id} may only be used once.\cr
If \code{NULL} IDs are an integer sequence from one to the number of rows.
Default is \code{NULL}.}
}
\value{
(\code{data.frame})\cr
Agents \code{data.frame} with 7 columns and as many rows as the longest vector passed as an argument
to this function. Column names are standardized.
}
\description{
Generates a data.frame to define Agents by their properties in a standardized format.
The output can be combined with the output of \code{\link[=create_studies]{create_studies()}} to define a \link{combine_agents_studies}(complete study).
}
\details{
As a convenience functionality, this function can take atomic vectors of equal length or length 1 as input. If the input is of length 1,
the value is recycled to match the length of the longest vector.
For more complex automatically generated inputs, users may want to use functions such as \code{\link[=rep]{rep()}} or \code{\link[=seq]{seq()}}.
}
\examples{
# Generate one agent, ID is generated automatically
create_agents(
  subj_effect_mu = 0.3, subj_effect_sigma = 1.2,
  subj_prob_fault_alpha = 0.2, subj_prob_fault_beta = 0.4,
  subj_error_size_mu = 0.6, subj_error_size_sigma = 0.2
)
# Generate multiple agents
create_agents(
  agent_id = c("Alice", "Bob"),
  subj_effect_mu = c(0.3, 0.4), subj_effect_sigma = c(1.2, 1.3),
  subj_prob_fault_alpha = c(0.2, 0.3), subj_prob_fault_beta = c(0.4, 0.5),
  subj_error_size_mu = c(0.6, 0.7), subj_error_size_sigma = c(1.2, 1.3)
)

}
