% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents_studies.R
\name{create_studies}
\alias{create_studies}
\title{Create data.frame of Studies with all Required Properties}
\usage{
create_studies(
  N,
  resources,
  cost,
  benefit,
  obj_prob_fault,
  obj_effect_mu,
  obj_effect_sigma2,
  obj_error_size_mu,
  obj_error_size_sigma2,
  study_id = NULL,
  agent_id = NULL
)
}
\arguments{
\item{N}{(\code{integer()})\cr
The number of code units in the study.}

\item{resources}{(\code{numeric()})\cr
The total resources aviailable for the study.}

\item{cost}{(\code{numeric()})\cr
The cost associated with one search round for the particular study.}

\item{benefit}{(\code{numeric()})\cr
The benefit associated with one search round for the particular study.}

\item{obj_prob_fault}{(\code{numeric()})\cr
The probability of making an error in a search round (Bernoulli).}

\item{obj_effect_mu}{(\code{numeric()})\cr}

\item{obj_effect_sigma2}{(\code{numeric()})\cr}

\item{obj_error_size_mu}{(\code{numeric()})\cr
The expected value of the objective error size distribution (normal).}

\item{obj_error_size_sigma2}{(\code{numeric()})\cr
The variance of the objective error size distribution (normal).}

\item{study_id}{(\code{character()} or \code{integer()})\cr
Identifiers for Studies, has to be \strong{unique}, i.e. the same \code{study_id} may only be used once.\cr
If \code{NULL} IDs are an integer sequence from one to the number of rows. Default is \code{NULL}.}

\item{agent_id}{(\code{character()} or \code{integer()})\cr
Identifiers for Agents, has to be \strong{unique}, i.e. the same \code{agent_id} may only be used once.\cr
If \code{NULL} IDs are an integer sequence from one to the number of rows. Default is \code{NULL}.}
}
\value{
Studies \code{data.frame} with 11 columns and as many rows as the longest vector passed as an argument.
Column names are standardized.
}
\description{
Generates a data.frame to define Studies by their properties in a standardized format.
The output can be combined with the output of \code{\link[=create_agents]{create_agents()}} to define a \link[=combine_agents_studies]{complete study}.
}
\details{
As a convenience functionality, this function can take atomic vectors of equal length or length 1 as input. If the input is of length 1,
the value is recycled to match the length of the longest vector.
For more complex automatically generated inputs, users may want to use functions such as \code{\link[=rep]{rep()}} or \code{\link[=seq]{seq()}}.
}
\examples{
# Create a study
study <- create_studies(
  study_id = "Alice2022", agent_id = "Alice",
  N = 30, resources = 1000, cost = 400, benefit = 20,
  obj_prob_fault = 0.4,
  obj_effect_mu = 0.4, obj_effect_sigma2 = 0.5,
  obj_error_size_mu = 0.1, obj_error_size_sigma2 = 0.2
)
# Create multiple studies
studies <- create_studies(
  study_id = c("Alice2022", "BobEtAl2024"), agent_id = c("Alice", "Bob"),
  N = c(30, 20), resources = 100, cost = c(5, 10), benefit = c(10, 15),
  obj_prob_fault = 0.4,
  obj_effect_mu = 0.4, obj_effect_sigma2 = 0.5,
  obj_error_size_mu = 0.1, obj_error_size_sigma2 = 0.2
)
}
