% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_distributions_heatmap}
\alias{plot_distributions_heatmap}
\title{Plot Probability Distributions over Rounds as Heatmap}
\usage{
plot_distributions_heatmap(literature, study_id, type)
}
\arguments{
\item{literature}{(\code{list()})\cr
A list containing the simulated literature, generated by \code{\link[=simulate_literature]{simulate_literature()}}.}

\item{study_id}{(\code{character()} or \code{integer()})\cr
A unique identifier of a study indicating for which study the plot should be generated.}

\item{type}{(\code{character()})\cr
a character string indicating the type of distribution to plot. Can be either "after_effect" or "after_fault_ind".}
}
\value{
A \code{ggplot} object that can be further customized using the \code{ggplot2} package.
}
\description{
Plots the discrete probability distributions simulated through \code{\link[=simulate_literature]{simulate_literature()}} per round as a heatmap.
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
library(ggplot2)
# Create study and agent
agent <- create_agents(
  agent_id = "Alice",
  subj_effect_mu = 0.3, subj_effect_sigma = 1.2,
  subj_prob_fault_alpha = 0.2, subj_prob_fault_beta = 0.4,
  subj_error_size_mu = 0.6, subj_error_size_sigma = 0.2
)
study <- create_studies(
  study_id = "Alice2022", agent_id = "Alice",
  N = 30, resources = 1000, cost = 400, benefit = 20,
  obj_effect_mu = 0.4, obj_effect_sigma = 0.5,
  obj_prob_fault = 0.4,
  obj_error_size_mu = 0.1, obj_error_size_sigma = 0.2
)
# Simulate the study
lit <- simulate_literature(agents = agent, studies = study, seed = 123)
# Plot the probability distributions of remaining faults after the
# effect size was observed in the respective round
plot_distributions_heatmap(lit, "Alice2022", "after_effect")
}

}
