% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_histogram_final_effect_sizes}
\alias{plot_histogram_final_effect_sizes}
\title{Plot Histogram of Final Observed Effect Sizes}
\usage{
plot_histogram_final_effect_sizes(literature, binwidth = 0.3)
}
\arguments{
\item{literature}{(\code{list})\cr
The simulated literature data, generated by \code{\link[=simulate_literature]{simulate_literature()}}.}

\item{binwidth}{(\code{numeric(1)})\cr
The width of the histogram bins.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot Histogram of Final Observed Effect Sizes
}
