% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_obs_effect_sizes}
\alias{plot_obs_effect_sizes}
\title{Plot Observed Effect Sizes per Round}
\usage{
plot_obs_effect_sizes(literature, study_id = NULL, true_effect = TRUE)
}
\arguments{
\item{literature}{(\code{list()})\cr
A list containing the simulated literature, generated by \code{\link[=simulate_literature]{simulate_literature()}}.}

\item{study_id}{(\code{character()} or \code{integer()})\cr
A vector of (possibly multiple) unique identifiers of studies indicating for which studies the plot should be generated.
If \code{NULL}, all studies in \code{literature} are plotted. Default is \code{NULL}.}

\item{true_effect}{(\code{logical()})\cr
If \code{TRUE}, the true effect sizes of each study will be plotted additionally. Default is \code{TRUE}.}
}
\value{
A \code{ggplot} object that can be further customized using the \code{ggplot2} package.
}
\description{
Plot Observed Effect Sizes per Round
}
