% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_summary_panel}
\alias{plot_summary_panel}
\title{Plot Summary Panel}
\usage{
plot_summary_panel(
  literature,
  study_id = NULL,
  cbr = FALSE,
  true_effect = FALSE
)
}
\arguments{
\item{literature}{(\code{list()})\cr
A list containing the simulated literature, generated by \code{\link[=simulate_literature]{simulate_literature()}}.}

\item{study_id}{(\code{character()} or \code{integer()})\cr
A vector of (possibly multiple) unique identifiers of studies indicating for which studies the plot should be generated.
If \code{NULL}, all studies in \code{literature} are plotted. Default is \code{NULL}.}

\item{cbr}{(\code{logical()})\cr
If \code{TRUE}, the cost-benefit-ratios will be plotted additionally in the first plot. Default is \code{FALSE}.}

\item{true_effect}{(\code{logical()})\cr
If \code{TRUE}, the true effect sizes of each study will be plotted additionally in the third plot. Default is \code{FALSE}.}
}
\value{
A \code{patchwork} / \code{gg} object that can be further modified using the \code{patchwork} package.
}
\description{
Plots a panel of three plots: the probability of finding a fault per round,
the percentage of remaining faults per round, and the observed effect sizes per round.
}
\examples{
# Create study and agent
agent <- create_agents(
  agent_id = "Alice",
  subj_effect_mu = 0.3, subj_effect_sigma = 1.2,
  subj_prob_fault_alpha = 0.2, subj_prob_fault_beta = 0.4,
  subj_error_size_mu = 0.6, subj_error_size_sigma = 0.2
)
study <- create_studies(
  study_id = "Alice2022", agent_id = "Alice",
  N = 30, resources = 1000, cost = 4, benefit = 20,
  obj_effect_mu = 0.4, obj_effect_sigma = 0.5,
  obj_prob_fault = 0.4,
  obj_error_size_mu = 0.1, obj_error_size_sigma = 0.2
)
# Simulate the study
lit <- simulate_literature(agents = agent, studies = study, seed = 123)
# Plot the summary panel
plot_summary_panel(lit)

}
