% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scerrModel-package.R
\docType{package}
\name{scerrModel-package}
\alias{scerrModel}
\alias{scerrModel-package}
\title{scerrModel: A Formal Model for Scientific Errors in Data Analysis}
\description{
This package implements a formal model of scientific errors in data analysis based on the work from Kohrt et al. (2023) and Melinscak (2023).
}
\author{
\strong{Maintainer}: Keno Mersmann \email{keno.mersmann@campus.lmu.de}

}
\keyword{internal}
