% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_literature.R
\name{simulate_literature}
\alias{simulate_literature}
\title{Simulation of literature data}
\usage{
simulate_literature(
  complete_studies,
  agents = NULL,
  studies = NULL,
  seed = NULL,
  keep_seed_const = FALSE
)
}
\arguments{
\item{complete_studies}{(\code{data.frame})\cr
A \code{data.frame} built with \code{\link[=combine_agents_studies]{combine_agents_studies()}} containing all properties of the agents and studies that are to be simulated.}

\item{agents}{(\code{data.frame})\cr
A \code{data.frame} built with \code{\link[=create_agents]{create_agents()}} containing the agents' properties. If \code{complete_studies} is given, this parameter is ignored.
Otherwise this is combined with \code{studies} to create \code{complete_studies}.}

\item{studies}{(\code{data.frame})\cr
A \code{data.frame} built with \code{\link[=create_studies]{create_studies()}} containing the studies' properties. If \code{complete_studies} is given, this parameter is ignored.
Otherwise this is combined with \code{agents} to create \code{complete_studies}.}

\item{seed}{(\code{integer()})\cr
The seed used for simulation. See Details for more information.
If \code{NULL}, the seed is not set. Default is \code{NULL}.}

\item{keep_seed_const}{(\code{logical()})\cr
If \code{TRUE}, the seed is kept constant for all studies. See Details for more information.
Default is \code{FALSE}.}
}
\value{
(named \code{list})
A list containing the input (\code{complete_studies} (passed or generated with \code{agents} and \code{studies}), \code{seed}, \code{keep_seed_const}) as well as
the simulation output as named lists (named as the respective \code{study_id}) for each study, with the following elements:
\itemize{
\item \code{obj_effect_size} (\code{numeric(1)})\cr
The simulated effect size in the objective reality.
\item \code{fault_indicators} (\code{numeric(N)})\cr
The simulated fault indicators per code unit.
\item \code{error_sizes} (\code{numeric(N)})\cr
The simulated error sizes per code unit.
\item \code{observed_effect_sizes} (\code{numeric(N+1)})\cr
The simulated effect sizes the agent might observe. The last entry is always equal to \code{obj_effect_size}.
\item \code{p_after_fault_ind} (\code{matrix(N+1, N+1)})\cr
Matrix of the discrete probability distribution of the number of remaining faults after observing the fault indicator.\cr
Rows represent the search iteration, columns the number of remaining faults, where column one represents zero remaining faults
and column \code{N+1} represents \code{N} remaining faults. \cr
One row contains probabilities of all realizations of a discrete probability distribution.
Row one is initialized with the prior belief about the number of remaining faults (Beta-Binomial(\code{N}, \code{subj_prob_fault_alpha}, \code{subj_prob_fault_beta})).
and row \code{N+1} shows the distribution after the last search round.
\item \code{p_after_effect} (\code{matrix(N, N+1)})\cr
Matrix of the discrete probability distribution of the number of remaining faults after observing the current effect size.\cr
Rows represent the search iteration, columns the number of remaining faults, where column one represents zero remaining faults
and column \code{N+1} represents \code{N} remaining faults. \cr
One row contains probabilities of all realizations of a discrete probability distribution.
\item \code{stopped_in_round} (\code{integer(1)})\cr
The round in which the agent stopped searching. If \code{N+1}, the agent finished the last round successfully (instead of stopping in the last round.).
\item \code{stopping_reason} (\code{character(1)})\cr
The reason for stopping the search. Can be "Search completed", "Expected utility too low" or "Resources depleted".
\item \code{p_fault_this_round} (\code{numeric(N)})\cr
The subjective probability of observing a fault in the respective search round. Used to calculate \code{eu_criterion}.
\item \code{eu_criterion} (\code{numeric(N)})\cr
The expected utility criterion for continuing to search in the respective search round. Used to decide whether to continue searching or not.
\item \code{remaining_resources} (\code{numeric(1)})\cr
The remaining resources at the point the agent decided to stop searching.
}
}
\description{
Simulates a literature data set based on the given parameters describing the studies.
}
\details{
The user may \code{\link[=set.seed]{set.seed()}} before calling this function to ensure reproducibility of the simulation. However, setting the arguments \code{seed} and additionally \code{keep_seed_const} can offer
some additional control.
If \code{keep_seed_const = FALSE} (default), the simulation uses all seeds from \code{seed} to \code{seed + nrow(complete_studies)} to keep the results of different studies independent while enabling the user
to change properties of one study without affecting the results of the simulation of other studies.
If performing multiple separate simulations, to ensure generation of independent literature data, use seeds that are further apart.
If \code{keep_seed_const = TRUE}, the seed is kept constant for all studies. This can be used to analyse differences in the simulation output due to changes in the study parameters while keeping the \code{seed} constant.
Be aware that simulations will be dependent.
}
