% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{simulation_summary}
\alias{simulation_summary}
\title{Summarize Simulated Literature Data}
\usage{
simulation_summary(literature, simple = TRUE)
}
\arguments{
\item{literature}{(\code{list})\cr
The simulated literature data, generated by \link{simulate_literature}.}

\item{simple}{(\code{logical()})\cr
If \code{TRUE}, the output will be simplified to only contain a selection of key columns.
If \code{FALSE} the output will also contain all columns of \code{complete_studies} passed to / created by \link{simulate_literature}.
Default is \code{TRUE}.}
}
\value{
(\code{data.frame})\cr
A \code{data.frame} giving an overview over the simulated literature, with the following columns if \code{simple = FALSE}:
\itemize{
\item \code{study_id} (\code{integer()} or \code{character()})\cr
The unique identifier of the study
\item \code{agent_id} (\code{integer()} or \code{character()})\cr
The unique identifier of the agent who performed the data analysis.
\item \code{N} (\code{integer()})\cr
Number of code units in the study.
\item \code{resources} (\code{numeric()})\cr
Total resources available for the study.
\item \code{cost} (\code{numeric()})\cr
Cost associated with one search round for the study.
\item \code{benefit} (\code{numeric()})\cr
Benefit associated with one search round for the study.
\item \code{bcr} (\code{numeric()})\cr
Benefit-cost ratio for the study.
\item \code{n_rounds} (\code{integer()})\cr
Number of rounds the agent searched for faults.
If \code{N+1}, this means that the agent finished the last round instead of stopping in the last round.
\item \code{reason} (\code{character()})\cr
Reason for stopping the search. Can be "Search completed", "Expected utility too low" or "Resources depleted".
\item \code{n_faults_total} (\code{integer()})\cr
Total number of faults in the study.
\item \code{n_faults_discovered} (\code{integer()})\cr
Number of faults discovered by the agent.
\item \code{n_faults_undiscovered} (\code{integer()})\cr
Number of faults not discovered by the agent.
\item \code{real_effect_sizes} (\code{numeric()})\cr
Real effect size of the study (i.e. in the objective reality).
\item \code{final_observed_effect_sizes} (\code{numeric()})\cr
Final observed effect size of the study in the round the agent stopped searching.
If \code{simple = TRUE} the returned \code{data.frame} contains all columns listed above as well as
all columns of \code{complete_studies} passed to / created by \link{simulate_literature}.
}
}
\description{
This function summarizes the simulated literature returned by \link{simulate_literature}.
}
